fix(extension-positioner): `Positioner` as value
