+++ .changeset/hungry-gifts-watch.md
@@ -1,0 +1,10 @@
+---
+'@remirror/extension-positioner': minor
+'remirror': patch
+---
+
+Export `Positioner` as a value.
+
+- Enable creating a positioner from another positioner with `Positioner.fromPositioner`
+
+Closes #441
