+++ packages/@remirror/extension-positioner/src/index.ts
@@ -8,16 +8,23 @@ export { getPositioner, PositionerExtension } from './positioner-extension';
 
 export { hasStateChanged, isEmptyBlockNode } from './positioner-utils';
 
-export type {
-  ElementsAddedParameter,
-  GetPositionParameter,
-  Positioner,
-  VirtualPosition,
-  VirtualNode,
-} from './positioners';
 export {
   centeredSelectionPositioner,
+  cursorPopupPositioner,
   emptyVirtualPosition,
   floatingSelectionPositioner,
-  cursorPopupPositioner as noSelectionPopupMenu,
 } from './positioners';
+
+export type {
+  GetPositionParameter,
+  VirtualPosition,
+  GetActiveParameter,
+  ElementsAddedParameter,
+  VirtualNode,
+  Coords,
+  BasePositioner,
+  BasePositionerParameter,
+  PositionerUpdateEvent,
+  SetActiveElement,
+} from './positioner';
+export { Positioner } from './positioner';
